+++ tools/bot/actions/MapServerMini.h
@@ -111,7 +111,7 @@ public:
             std::vector<Content> contentList;*/
             BlockObject *blockObject;
         };
-        uint8_t *map;//0x00 is not accessible, it's why don't have layer for it
+        uint16_t *map;//0x00 is not accessible, it's why don't have layer for it
         std::vector<Layer> layers;//layer 1=index 0, layer 2=index 1, ...
     };
     struct BlockObjectPathFinding{
